+++ chai/timeMan.h
@@ -6,4 +6,4 @@ int remainingHalfMoves(board_t* b);
 
 int allocateTime(board_t* b, int timeLeft, int inc);
 
-bool isTimeLeft(search_t* s);
+bool isTimeLeft(instr_t* instr);