+++ chai/uci.cpp
@@ -1,10 +1,29 @@
 #include "uci.h"
 
+
 static bool strStartsWith(std::string str, std::string start) {
 	return str.rfind(start, 0) == 0;
 }
 
-void uciMode(board_t* b, search_t* s) {
+void init() {
+	initClearSetMask();
+	initSquareToRankFile();
+	initAttackerMasks();
+	initMVV_LVA();
+	initEvalMasks();
+	initManhattenMask();
+
+	initObstructed();
+	initLine();
+	initSearch();
+
+	initRookMasks();
+	initRookMagicTable();
+	initBishopMasks();
+	initBishopMagicTable();
+}
+
+void uciMode(board_t* b, stats_t* s, instr_t* i) {
 	std::string cmd;
 
 	printUCI_Info();
@@ -13,34 +32,13 @@ void uciMode(board_t* b, search_t* s) {
 		getline(std::cin, cmd);
 		fflush(stdout);
 
-		if (!cmd.compare("quit")) {
-			break;
-		}
+		if (!cmd.compare("quit")) break;
+		if (!cmd.compare("uci")) printUCI_Info();
+		if (!cmd.compare("isready")) cout << "readyok" << endl;
+		if (!cmd.compare("ucinewgame")) uciParsePosition(b, "position startpos");
 
-		if (!cmd.compare("uci")) {
-			printUCI_Info();
-		}
- 
-		if (!cmd.compare("isready")) {
-			init();
-			initHashTables();
-
-			uciParsePosition(b, "position startpos");
-			cout << "readyok\n";
-		}
-
-		if (!cmd.compare("ucinewgame")) {
-			uciParsePosition(b, "position startpos");
-		}
-
-		if (strStartsWith(cmd, "position")) {
-			uciParsePosition(b, cmd);
-		}
-
-		if (strStartsWith(cmd, "go")) {
-			uciParseGo(b, s, cmd);
-		}
-
+		if (strStartsWith(cmd, "position")) uciParsePosition(b, cmd);
+		if (strStartsWith(cmd, "go")) uciParseGo(b, s, i, cmd);
 		if (strStartsWith(cmd, "setoption")) {
  			try {
 				uciSetOption(cmd);
@@ -51,37 +49,39 @@ void uciMode(board_t* b, search_t* s) {
 
 		fflush(stdout);
 	}
-
-	cout << "Left uci mode." << endl;
-
 }
 
 void uciSetOption(std::string cmd) {
 
 	if (strStartsWith(cmd, "setoption name Hash value ")) {
-		int newMbSize = stoi(cmd.substr(strlen("setoption name Hash value "), std::string::npos));
-		if (resizeHashTables(newMbSize)) {
-			cout << "info string set Hash to " << newMbSize << "MB" << endl;
+		int newMbSize = stoi(cmd.substr(26, std::string::npos));
+
+		if (!resizeHashTables(newMbSize)) {
+			cerr << "info string Requested hash size is not inside ["
+				 << MIN_TT_SIZE << ", " << MAX_TT_SIZE << "]" << endl;
+			return;
 		}
+
+		cout << "info string set Hash to " << newMbSize << "MB" << endl;
 	}
 
 	if (strStartsWith(cmd, "setoption name SyzygyPath value ")) {
-		std::string syzygyPath = cmd.substr(strlen("setoption name SyzygyPath value "), std::string::npos);
-		// freeEGTB(); error if used to reset path
+		std::string syzygyPath = cmd.substr(32, std::string::npos);
+
         if (!syzygyPath.compare("")) {
-        	cout << "info string Error: SyzygyPath is empty." << endl;
+        	cout << "info string SyzygyPath is empty." << endl;
 			return;
 		}
+
 		initEGTB(syzygyPath.c_str());
-        cout << "info string set SyzygyPath to " 
-			 << syzygyPath << ". Max TB=" << TB_LARGEST << endl;
+        cout << "info string set SyzygyPath to " << syzygyPath << ". Max TB=" << TB_LARGEST << endl;
     }
 
 	if (strStartsWith(cmd, "setoption name Threads value ")) {
-		int requestedThreads = stoi(cmd.substr(strlen("setoption name Threads value "), std::string::npos));
+		int requestedThreads = stoi(cmd.substr(29, std::string::npos));
 		
 		if (!resizeThreadPool(requestedThreads)) {
-			cout << "info string Threads value is invalid." << endl;
+			cout << "info string Requested Thread value is invalid." << endl;
 			return;
 		}
 
@@ -125,83 +125,13 @@ void uciParsePosition(board_t* b, std::string cmd) {
 
 	fflush(stdout);
 }
-/*
-void uciParseGo(board_t* b, search_t* s, std::string cmd) {
-	s->startTime = getTimeMs();
 
-	int depth = -1;
-
-	// int movesLeft = 30;
-	int movesLeft = -1;
-
-	int moveTime = -1;
-	int time = -1;
-	int inc = 0;
-	s->timeSet = false;
-
-	std::istringstream iss(cmd);
-	std::vector<std::string> tokens{ std::istream_iterator<std::string>{iss},
-					 std::istream_iterator<std::string>{} };
-
-	for (int i = 0; i < (int)tokens.size() - 1; i++) {
-		if (tokens[i] == "infinite") continue;
-		if (tokens[i] == "binc" && b->stm == chai::BLACK) inc = stoi(tokens[i + 1]);
-		if (tokens[i] == "winc" && b->stm == chai::WHITE) inc = stoi(tokens[i + 1]);
-		if (tokens[i] == "btime" && b->stm == chai::BLACK) time = stoi(tokens[i + 1]);
-		if (tokens[i] == "wtime" && b->stm == chai::WHITE) time = stoi(tokens[i + 1]);
-		if (tokens[i] == "movestogo") movesLeft = stoi(tokens[i + 1]);
-		if (tokens[i] == "movetime") moveTime = stoi(tokens[i + 1]);
-		if (tokens[i] == "depth") depth = stoi(tokens[i + 1]);
-	}
-
-	// if fixed search time is set
-	if (moveTime != -1) {
-		time = moveTime;
-		movesLeft = 1;
-	}
-
-	//cout << "Parsed wbtime:" << time << " wbinc:" << inc << " movetime:" << moveTime << endl;
-
-	// if stop is set
-	// s->startTime = getTimeMs();
-	if (time != -1) {
-		// if (movesLeft < 1 || movesLeft > 30) movesLeft = 30;
-		movesLeft = (movesLeft == -1) ? remainingHalfMoves(b)
-									  : movesLeft;
-
-		s->timeSet = true;
-
-		// catch low time (ok for tc 20/0.3)
-		if (time <= 3000) {
-			s->stopTime = s->startTime + (inc / 8);
-		} else if (time <= 4000) {
-			s->stopTime = s->startTime + (inc / 4);
-		} else {
-			time /= movesLeft;
-			time -= 50;
-			//s->stopTime = s->startTime + time + inc;
-			s->stopTime = s->startTime + time;
-		}
-
-		Assert(s->startTime < s->stopTime);
-	}
-
-	if (depth == -1) {
-		s->depth = MAX_DEPTH;
-	} else {
-		s->depth = depth;
-	}
-
-	search(b, s);
-}
-*/
-
-void uciParseGo(board_t* b, search_t* s, std::string cmd) {
-	s->startTime = getTimeMs();
+void uciParseGo(board_t* b, stats_t* s, instr_t* instr, std::string cmd) {
+	instr->startTime = getTimeMs();
 
 	int depth         = 0; // search to this depth
 	int timeLeft      = 0; // time left on (our) clock
-	int inc;  			// increment
+	int inc;			   // increment
 
 	std::istringstream iss(cmd);
 	std::vector<std::string> tokens{ std::istream_iterator<std::string>{iss},
@@ -216,50 +146,17 @@ void uciParseGo(board_t* b, search_t* s, std::string cmd) {
 		if (tokens[i] == "depth") depth = stoi(tokens[i + 1]);
 	}
 
-	// Depth is either set as go depth command or MAX_DEPTH of engine
-	s->depth = (depth != 0) ? depth : MAX_DEPTH;
+	// Depth is either set as "go depth x" command or MAX_DEPTH of engine
+	instr->depth = (depth != 0) ? depth : MAX_DEPTH;
 
 	// If timeLeft is provided, search with time constraints
+	instr->timeSet = false;
 	if (timeLeft) {
-		s->timeSet       = true;
-		s->timeLeft      = timeLeft;
-		s->allocatedTime = allocateTime(b, timeLeft, inc);
+		instr->timeSet       = true;
+		instr->timeLeft      = timeLeft;
+		instr->allocatedTime = allocateTime(b, timeLeft, inc);
 	}
 
-	search(b, s);
-}
-
-void init() {
-#ifdef INFO
-	auto start = std::chrono::high_resolution_clock::now();
-#endif //INFO
-	initClearSetMask();
-	initSquareToRankFile();
-	initAttackerMasks();
-	initMVV_LVA();
-	initEvalMasks();
-	initManhattenMask();
-
-	initObstructed();
-	initLine();
-	initSearch();
-
-#ifdef INFO
-	auto stop = std::chrono::high_resolution_clock::now();
-	auto duration = std::chrono::duration_cast<std::chrono::milliseconds>(stop - start);
-	cout << "Init keys and masks ... " << duration.count() << "ms\n";
-	start = std::chrono::high_resolution_clock::now();
-#endif // !INFO
-
-	initRookMasks();
-	initRookMagicTable();
-	initBishopMasks();
-	initBishopMagicTable();
-
-#ifdef INFO
-	stop = std::chrono::high_resolution_clock::now();
-	duration = std::chrono::duration_cast<std::chrono::milliseconds>(stop - start);
-	cout << "Init magic tables for bishop and rooks ... " << duration.count() << "ms\n" << endl;
-#endif // !INFO
+	search(b, s, instr);
 }