+++ chai/uci.h
@@ -14,9 +14,9 @@
  * Enters UCI mode. Described in http://wbec-ridderkerk.nl/html/UCIProtocol.html
  *
  * @param  b A board_t to process.
- * @param  s A search_t to process.
+ * @param  s A stats_t to process.
  */
-void uciMode(board_t* b, search_t* s);
+void uciMode(board_t* b, stats_t* s, instr_t* i);
 
 void uciSetOption(std::string cmd);
 
@@ -32,10 +32,10 @@ void uciParsePosition(board_t* b, std::string cmd);
  * Parse search information from UCI go command.
  *
  * @param  b   A board_t to process.
- * @param  s   A search_t to process.
+ * @param  s   A stats_t to process.
  * @param  cmd The command.
  */
-void uciParseGo(board_t* b, search_t* s, std::string cmd);
+void uciParseGo(board_t* b, stats_t* s, instr_t* instr, std::string cmd);
 
 /**
  * Complete initialization for chess engine. Called once at startup.